dev build 1.1b1
