+++ app/src/App.js
@@ -15,29 +15,6 @@ export default function App() {
 					</Route>
 				</Routes>
 			</BrowserRouter>
-			<Helmet>
-				<script>
-					{`
-					let tag;
-					let player = null;
-					let serverResponse = false;
-
-					window.setServerResponse = (value) => {
-						serverResponse = value;
-					}
-
-					window.getPlayer = () => {
-						return player;
-					}
-
-					setInterval(() => {
-						fetch('${process.env.REACT_APP_SERVER_URL}/keepalive', {
-							method: 'get'
-						})
-					}, 300000);
-					`}
-				</script>
-			</Helmet>
 		</>
 	)
 }
