+++ app/src/utils/context.js
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+
+export const SearchContext = createContext(null);
+export const VideoContext = createContext(null);
